+++ src/gui/button.cpp
@@ -2,7 +2,7 @@
 
 using namespace gui;
 
-Button::Button(WidgetCollection *parent, i32 x, i32 y, i8 w, i8 h, u8 r, u8 g, u8 b) 
+Button::Button(WidgetCollection *parent, i16 x, i16 y, i16 w, i16 h, u8 r, u8 g, u8 b) 
   : Widget(parent, x, y, w, h, r, g, b) {
   down = false;
 }
@@ -17,7 +17,7 @@ void Button::OnSetup() {
   context->FillRectangle(x + down, y + down, w - (down * 2), h - (down * 2), r, g, b);
 }
 
-void Button::OnMouseMove(i32 oldX, i32 oldY, i32 x, i32 y) {
+void Button::OnMouseMove(i16 oldX, i16 oldY, i16 x, i16 y) {
   if(down && (
       oldX == this->x || 
       oldX == w + this->x || 
@@ -30,7 +30,7 @@ void Button::OnMouseMove(i32 oldX, i32 oldY, i32 x, i32 y) {
   }
 }
 
-void Button::OnMouseDown(i32 x, i32 y, u8 button) {
+void Button::OnMouseDown(i16 x, i16 y, u8 button) {
   if(button == 1) {
     down = true;
     context->FillRectangle(this->x, this->y, w, 1, ~r, ~g, ~b);
@@ -40,7 +40,7 @@ void Button::OnMouseDown(i32 x, i32 y, u8 button) {
   }
 }
 
-void Button::OnMouseUp(i32 x, i32 y, u8 button) {
+void Button::OnMouseUp(i16 x, i16 y, u8 button) {
   if(button == 1) {
     down = false;
     context->FillRectangle(this->x, this->y, w, 1, r, g, b);
